+++ src/pygama/dsp/processors/get_multi_time_pickoff.py
