+++ src/pygama/dsp/processors/sipm_noise_regulation.py
